Inline css & js to make preview work
YES! This was probably one of the biggest hurdles to take.
The iframe that shows the preview can not load relative sources.
If it does load relative sources, it would get them from the live site.
But those sources could be changed inside the web browser fs.
But there is no way to serve the sources inside the web browser fs,
because the requests would go directly to the server.
Therefor all relative sources are inlined inside the html.