Add template search and install (WIP)
Big commit. Adds a new drawer as a separated component.
TODO: Separate the other drawers like this as well.
Added navigation store to keep state of the available drawers, and the
current active drawer.
The functionailty to run metalsmith has now been moved to this drawer.
Added functionality to search for repo's with tag `neumannssg-template`.
Added functionailty to add a template as a submodule to .gitmodules.
.gitmodules will be created if it wasn't already.
TODO: Add link to submodule, create commit.
See: <https://stackoverflow.com/questions/35765445/how-to-create-submodule-in-github-using-github-rest-api/>
TODO: Make sure GetFile works with files in a submodule.