Refactor openFiles to use fileContents
openFiles is now a getter in the github store.
It returns the files in fileContents with a property opened that is true.

The watchers for `active` and `openTab` now interact with fileContents
directly. fileContents is mapped using mapState.
An opened file is marked as such by giving it a property opened is true.

Te `closeTab` method removes the opened property using the mutation
`setFileOpened`. It then selects the right tab to show.
This selecting algorithm is no simplified but yields the same results.

Github store is now aware of the file's encoding.
This means non-binary files get converted to utf-8.
The encoding is shown through a property `encoding`.
No property means the decoding is `base64`, the default.
Upon creating blobs the utf-8 files are converted to base64.

For safe encoding BestBase64EncoderDecoder package is used.
The original plugin used for encoding is removed.