Fix treeview for long filenames
This commit truncates the labels in treeview.
Long filenames where pushing the folder buttons out of sight.
The buttons are also moved to the append slot, which is the proper place.
But I didn't get it to work properly before.
A tooltip is added so that the full file name is still visible on hover.