POC JSPM and SystemJS
I think I got Metalsmith plugins to work using the JSPM CDN.
https://jspm.org/docs/cdn

I couldn't get dynamic loading of plugins to work.
It is hard to know when jspm is done loading.
So I put all the scripts in the header, using nuxt.

After that metalsmith seems to run, but pagination returns an error
saying that collections.posts doesn't exist.

Tried to throw some debugging info with the debug plugin but that
doesn't seem to throw anything anymore. So it seems as though
plugin behavior does change a bit after SystemJS transpiles them.